+++ packages/common-utils/src/clickhouse/browser.ts
@@ -98,6 +98,7 @@ export class ClickhouseClient extends BaseClickhouseClient {
 
   protected async __query<Format extends DataFormat>({
     query,
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion -- default generic value
     format = 'JSON' as Format,
     query_params = {},
     abort_signal,
@@ -123,6 +124,7 @@ export class ClickhouseClient extends BaseClickhouseClient {
         : {}),
     };
 
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion -- client library type mismatch
     return this.getClient().query({
       query,
       query_params,